+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds the `sizeBasedOn` prop to the `Logo` component.
+
 ## 2.1.1 (October 26, 2023)
 
 ### Adds