+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds the `sizeBasedOn` prop to the `Logo` component.
 - Added the `autoComplete` prop to the `TextInput` component for setting the "autocomplete" attribute manually.
 
+### Updates
+
+- Updates max version for Chakra UI to 1.8.9 for minor bug fixes.
+
 ## 2.1.1 (October 26, 2023)
 
 ### Adds