+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+## 2.1.2 (November 9, 2023)
+
 ### Adds
 
 - Adds the `sizeBasedOn` prop to the `Logo` component.
@@ -15,7 +17,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Updates
 
-- Updates max version for Chakra UI to 1.8.9 for minor bug fixes.
+- Updates max version for Chakra UI to `1.8.9` for minor bug fixes.
+- Updates the layout for the `"campaign"` variant of the `Hero` component to have consistent padding on its left and right sides.
 
 ## 2.1.1 (October 26, 2023)